Feat: post rent returned /{id}/return

+++ src/main/java/com/boardcamp/api/dtos/RentalsDTO.java
@@ -17,5 +17,5 @@ public class RentalsDTO {
     
     @NotNull
     @Positive
-    private Long daysRented;
+    private Integer daysRented;
 }
